rename core -> context
