+++ context/src/main/scala/org/kalergic/contextual/v0/context/ContextValueListener.scala
@@ -1,0 +1,11 @@
+package org.kalergic.contextual.v0.context
+
+import scala.reflect.runtime.universe._
+
+trait ContextValueListener[V] {
+  def valueTypeTag: TypeTag[V]
+  def contextKey: ContextKey[V]
+  def name: String
+  def put(value: V): Unit
+  def removed(value: V): Unit
+}
