+++ context/src/main/scala/org/kalergic/contextual/v0/context/ShareableContext.scala
@@ -1,0 +1,7 @@
+package org.kalergic.contextual.v0.context
+
+private[context] trait ShareableContext extends Context {
+  private[context] def copy(): ShareableContext
+  private[context] def activateForCurrentThread(): Unit
+  private[context] def deactivateForCurrentThread(): Unit
+}
