--- core/src/main/scala/org/kalergic/contextual/v0/core/ShareableContext.scala
@@ -1,7 +1,0 @@
-package org.kalergic.contextual.v0.core
-
-private[core] trait ShareableContext extends Context {
-  private[core] def copy(): ShareableContext
-  private[core] def activateForCurrentThread(): Unit
-  private[core] def deactivateForCurrentThread(): Unit
-}
